bot/requirements.txt
aiohttp==3.9.5
aiomysql==0.2.0
aiosignal==1.3.1
alembic==1.13.3
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.6
discord.py==2.3.2
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
Mako==1.3.6
MarkupSafe==3.0.2
multidict==6.0.5
pycparser==2.22
PyMySQL==1.1.0
requests==2.31.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
urllib3==2.2.1
yarl==1.9.4
